Correct SPI Clock Configuration
From the recent Captures of the official firmware the SPI clock can be seen to be Active High triggering on the falling edge.
The OpenGD77 code was configured as active Low triggering on the leading edge.
This change corrects the setting.
Probably doesnt make any difference though as the actual clock edge is the same.